config

+++ config/experiment_config_multi_machine.py
@@ -26,9 +26,10 @@ def read_cfg(mod="non_contrastive"):
     FLAGS.loss_type = "symmetrized"  # asymmetrized (2 only options)
     # Moving average the weight From Online to Target Encoder Network
     # two options [fixed_value, schedule] schedule recommend from BYOL
+    
     FLAGS.moving_average = "schedule"
     # ['fp16', 'fp32'],  # fp32 is original precision
-    FLAGS.mixprecision = 'fp32'
+    FLAGS.mixprecision = 'fp16'
     FLAGS.precision_method = "API"  # "API"
     # , [ 'original', 'model_only', ],
     FLAGS.base_lr = 0.5
